--- a/queries.sql
+++ b/queries.sql
@@ -233,3 +233,9 @@ JOIN species ON species.id = animals.species_id
 GROUP BY species.name
 ORDER BY COUNT(visits.id)
 DESC LIMIT 1;
+
+EXPLAIN ANALYZE SELECT COUNT(*) FROM visits WHERE animal_id = 4;
+
+EXPLAIN ANALYZE SELECT * FROM visits where vet_id = 2;
+
+EXPLAIN ANALYZE SELECT * FROM owners where email = 'owner_18327@mail.com';