--- a/schema.sql
+++ b/schema.sql
@@ -64,3 +64,6 @@ CREATE TABLE visits (
   vet_id INT REFERENCES vets(id),
   date_of_visit DATE NOT NULL
 )
+
+-- Add an email column to your owners table
+ALTER TABLE owners ADD COLUMN email VARCHAR(120);