--- a/schema.sql
+++ b/schema.sql
@@ -67,3 +67,9 @@ CREATE TABLE visits (
 
 -- Add an email column to your owners table
 ALTER TABLE owners ADD COLUMN email VARCHAR(120);
+
+CREATE INDEX ix_visit_animal ON visits (animal_id);
+
+CREATE INDEX ix_visits ON visits (vet_id, id, animal_id, date_of_visit);
+
+CREATE INDEX ix_visits ON visits (vet_id, id, animal_id, date_of_visit);