# Vet clinic database: database performance audit
**The following changes were implemented:**

### On `README.md`:

- Added information about second author ✅ 
- Modified `Future Features` ✅ 

### On `LICENSE`:

- Added second author's name ✅ 

### On `schema.sql`:

- Added an `email` column to the `owners` table ✅ 
- Created an `index table` for the `visits` table's `animal_id` ✅ 
- Created an `index table` for the `visits` table ✅ 

### On `data.sql`:

- Inserted around **21,565,728** rows into the `visits` table ✅ 
- Inserted around **15,000,000** rows into the `owners` table ✅ 
- _**THE FOLLOWING SCREENSHOT DISPLAYS THE NEED FOR SEVERAL INSERTIONS OF DATA IN ORDER FOR US 

TO RUN AN `EXPLAIN ANALYZE` THAT COULD RETURN AN `EXECUTION TIME X >= 1000 MS`**_ 🤠 
![insert_data](https://github.com/mwismann/vet_clinic_2/assets/107666488/07388b26-1701-4ca0-816f-15ab9c972975)


### On `queries.sql`:

- **Ran a performance audit for the following queries:**

1. `EXPLAIN ANALYZE SELECT COUNT(*) FROM visits where animal_id = 4;`

- BEFORE: `*Execution Time X > 1300 ms*`
![q1_before](https://github.com/mwismann/vet_clinic_2/assets/107666488/257805fd-d36a-40b1-82da-a3d6d0acbf4a)

- AFTER: *`Execution Time X < 150 ms`* ✅
![q1_after](https://github.com/mwismann/vet_clinic_2/assets/107666488/36f6ce10-375a-4863-b0b9-21f100d591f9)


2. `EXPLAIN ANALYZE SELECT * FROM visits where vet_id = 2;`

- BEFORE: `*Execution Time X > 5600 ms*`
![q2_before](https://github.com/mwismann/vet_clinic_2/assets/107666488/dcc014f6-dd9e-4b62-b152-e4de202146f4)

- AFTER: *`Execution Time X < 2000 ms`* ✅ 
![q2_after](https://github.com/mwismann/vet_clinic_2/assets/107666488/1be07248-c966-4bab-9b93-7eda0b679a73)

3. `EXPLAIN ANALYZE SELECT * FROM owners where email = 'owner_18327@mail.com';`

- BEFORE: `*Execution Time X > 1400 ms*`
![q3_before](https://github.com/mwismann/vet_clinic_2/assets/107666488/6339df39-8cb8-4a2f-a512-d313dfbd84cd)

- AFTER: *`Execution Time X < 1 ms`* ✅ 
![q3_after](https://github.com/mwismann/vet_clinic_2/assets/107666488/63059283-c956-4616-8991-a382f5ef0da8)